fix: align Rust model types with DB schema (INT4/REAL)
basis_price and price changed from i64 to i32, discount_rate from f64
to f32 to match the PostgreSQL integer/real column types.

+++ src/model.rs
@@ -11,9 +11,9 @@ pub struct Model {
     pub amazon_url: String,
     pub kindle_id: Option<String>,
     pub title: String,
-    pub basis_price: Option<i64>,
-    pub price: Option<i64>,
-    pub discount_rate: Option<f64>,
+    pub basis_price: Option<i32>,
+    pub price: Option<i32>,
+    pub discount_rate: Option<f32>,
     pub updated_at: chrono::NaiveDateTime,
     pub active_at: Option<chrono::NaiveDateTime>,
 }
